requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.5
diff-match-patch==20241021
Django==5.2.3
django-import-export==4.3.8
et_xmlfile==2.0.0
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
